+++ boilerplate-simple-api/cmd/main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	hello_controller "simple-api/internal/modules/hello/controller"
+	hello_handler "simple-api/internal/modules/hello/handler"
 	pkg_http "simple-api/pkg/http"
 
 	"github.com/gin-gonic/gin"
@@ -13,7 +13,7 @@ const PORT int = 8000
 func main() {
 	srv := pkg_http.NewHTTPServer(gin.DebugMode)
 
-	hello_controller.BindHelloController(srv.Group("/hello"))
+	hello_handler.BindHelloHandler(srv.Group("/hello"))
 
 	srv.Run(fmt.Sprintf(":%d", PORT))
 }
