+++ boilerplate-simple-api/internal/modules/hello/handler/method_create_hello.go
@@ -1,4 +1,4 @@
-package hello_controller
+package hello_handler
 
 import (
 	"fmt"
@@ -9,9 +9,9 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-// path : /hello [PUT]
-func (hc *helloController) ReplaceHello(ctx *gin.Context) {
-	var req hello_request.ReplaceHello
+// path : /hello [POST]
+func (hc *helloController) CreateHello(ctx *gin.Context) {
+	var req hello_request.CreateHello
 	if err := ctx.ShouldBindJSON(&req); err != nil {
 		ctx.Error(err)
 		return
@@ -19,8 +19,6 @@ func (hc *helloController) ReplaceHello(ctx *gin.Context) {
 
 	ctx.JSON(
 		http.StatusOK,
-		pkg_http_wrapper.NewResponse(
-			fmt.Sprintf("Your name is replaced from %s to %s", req.CurrentName, req.NewName),
-		),
+		pkg_http_wrapper.NewResponseWithValue(fmt.Sprintf("Hello, %s you are %d years old", req.Name, req.Age), req),
 	)
 }
