+++ boilerplate-simple-api/internal/modules/hello/handler/method_create_hello_test.go
@@ -1,4 +1,4 @@
-package hello_controller_test
+package hello_handler_test
 
 import (
 	"bytes"
@@ -17,7 +17,7 @@ import (
 func TestCreateHello(t *testing.T) {
 	srv := setupTest()
 
-	t.Run("Create Hello Http Test 1", func(t *testing.T) {
+	t.Run("Successfully Sending Create Hello Request", func(t *testing.T) {
 		res := httptest.NewRecorder()
 		reqBody := hello_request.CreateHello{
 			Name: "Azie",
