+++ boilerplate-simple-api/internal/modules/hello/handler/method_delete_hello.go
@@ -1,4 +1,4 @@
-package hello_controller
+package hello_handler
 
 import (
 	"fmt"
