+++ boilerplate-simple-api/internal/modules/hello/handler/method_delete_hello_test.go
@@ -1,4 +1,4 @@
-package hello_controller_test
+package hello_handler_test
 
 import (
 	"encoding/json"
@@ -8,24 +8,28 @@ import (
 	pkg_http_wrapper "simple-api/pkg/http/wrapper"
 	"testing"
 
+	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestGetHelloByName(t *testing.T) {
+func TestDeleteHello(t *testing.T) {
 	srv := setupTest()
 
-	t.Run("Get Hello By Name Http Test 1", func(t *testing.T) {
+	t.Run("Successfully Sending Delete Hello Request", func(t *testing.T) {
 		res := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodGet, "/hello/azie", nil)
+
+		id := uuid.NewString()
+		req := httptest.NewRequest(http.MethodDelete, "/hello/"+id, nil)
 
 		srv.ServeHTTP(res, req)
 
-		name := "azie"
-		messageExpect := fmt.Sprintf("Hello, %s!", name)
+		messageExpect := fmt.Sprintf("Your Data by the id of %s is successfully deleted", id)
 
 		resBody := pkg_http_wrapper.NewResponse("")
 		err := json.Unmarshal(res.Body.Bytes(), &resBody)
 		assert.Nil(t, err)
+
+		assert.Empty(t, resBody.Error)
 
 		assert.Equal(t, messageExpect, resBody.Message)
 	})
