+++ boilerplate-simple-api/internal/modules/hello/handler/method_get_hello_by_name.go
@@ -1,4 +1,4 @@
-package hello_controller
+package hello_handler
 
 import (
 	"fmt"
