+++ boilerplate-simple-api/internal/modules/hello/handler/method_update_hello.go
@@ -1,4 +1,4 @@
-package hello_controller
+package hello_handler
 
 import (
 	"fmt"
@@ -9,9 +9,9 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-// path : /hello [POST]
-func (hc *helloController) CreateHello(ctx *gin.Context) {
-	var req hello_request.CreateHello
+// path : /hello [PATCH]
+func (hc *helloController) UpdateHello(ctx *gin.Context) {
+	var req hello_request.UpdateHello
 	if err := ctx.ShouldBindJSON(&req); err != nil {
 		ctx.Error(err)
 		return
@@ -19,6 +19,6 @@ func (hc *helloController) CreateHello(ctx *gin.Context) {
 
 	ctx.JSON(
 		http.StatusOK,
-		pkg_http_wrapper.NewResponseWithValue(fmt.Sprintf("Hello, %s you are %d years old", req.Name, req.Age), req),
+		pkg_http_wrapper.NewResponse(fmt.Sprintf("Your name is replaced to %s", req.NewName)),
 	)
 }
