+++ boilerplate-simple-api/internal/modules/hello/handler/type.go
@@ -1,4 +1,4 @@
-package hello_controller
+package hello_handler
 
 type helloController struct {
 }
